requirements.txt
numpy==1.16.2
opencv_python==3.4.8.29
Pillow==8.4.0
requests==2.27.1
setuptools==58.0.4
torch
torchvision
tqdm==4.62.3
face_recognition